refactor: 댓글 엔티티 관련 수정
